ci(macos): copy local dylib to package

+++ ci/macos/package-macos.sh
@@ -23,16 +23,23 @@ FILENAME="$PLUGIN_NAME-$PKG_VERSION.pkg"
 
 echo "=> Modifying $PLUGIN_NAME.so"
 mkdir lib
-for dylib in \
-	/usr/local/opt/cairo/lib/*.dylib \
-	/usr/local/opt/pango/lib/*.dylib \
-	/usr/local/opt/pangocairo/lib/*.dylib
-do
-	cp $dylib lib/
-	b=$(basename $dylib)
-	install_name_tool -id "@loader_path/$b" lib/$b
-	install_name_tool -change "$dylib" "@loader_path/../lib/$b" ./build/$PLUGIN_NAME.so
-done
+
+function copy_local_dylib
+{
+	local dylib
+	otool -L $1 | awk '/^	\/usr\/local\/(opt|Cellar)\/.*\.dylib/{print $1}' |
+	while read dylib; do
+		echo "Changing dependency $1 -> $dylib"
+		local b=$(basename $dylib)
+		if test ! -e lib/$b; then
+			cp $dylib lib/
+			chmod +rwx lib/$b
+			install_name_tool -id "@loader_path/$b" lib/$b
+			copy_local_dylib lib/$b
+		fi
+		install_name_tool -change "$dylib" "@loader_path/../lib/$b" $1
+	done
+}
 
 install_name_tool \
 	-change /tmp/obsdeps/lib/QtWidgets.framework/Versions/5/QtWidgets \
@@ -43,12 +50,15 @@ install_name_tool \
 		@executable_path/../Frameworks/QtCore.framework/Versions/5/QtCore \
 	./build/$PLUGIN_NAME.so
 
+copy_local_dylib ./build/${PLUGIN_NAME}.so
+
 # Check if replacement worked
 for dylib in ./build/$PLUGIN_NAME.so lib/*.dylib ; do
 	echo "=> Dependencies for $(basename $dylib)"
 	otool -L $dylib
 	echo "=> Search paths written in $(basename $dylib)"
 	otool -l $dylib
+	echo
 done
 
 if [[ "$RELEASE_MODE" == "True" ]]; then
